# api-gateway/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AeroLink Distributed Platform API",
    "version": "2.0.0",
    "description": "AeroLink Airline Systems Platform – Multi-Region Cloud API. Implements RESTful endpoints across 4 microservices: Booking (Aurora SQL), Baggage (DynamoDB NoSQL), Check-In, and Flight Operations. Secured with JWT/OAuth2 and RBAC.",
    "contact": { "name": "AeroLink Engineering", "email": "api@aerolink.io" },
    "license": { "name": "MIT" }
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local Development Gateway" },
    { "url": "https://api.aerolink.io", "description": "AWS Production (US-East-1)" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "schemas": {
...
      "BoardingPass": {
        "type": "object",
        "properties": {
          "checkinId": { "type": "string", "example": "CHK-12345" },
          "passenger": { "type": "string" },
          "flightId": { "type": "string" },
          "seat": { "type": "string", "example": "14A" },
          "gate": { "type": "string", "example": "B7" },
          "boardingTime": { "type": "string" },
          "issuedAt": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/bookings/all": {
      "get": {
        "summary": "List all bookings (Staff/Admin)",
        "tags": ["Booking Service"],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "All bookings joined with flight data" } }
      }
    },
    "/api/bookings/{id}": {
      "get": {
        "summary": "Get booking by ID",
        "tags": ["Booking Service"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "example": "BK-54321" }],
        "responses": {
          "200": { "description": "Booking details" },
          "404": { "description": "Booking not found" }
        }
      }
    },
...
    "/api/baggage/update": {
      "post": {
        "summary": "Update baggage status (Event-driven telemetry)",
        "description": "Performs a DynamoDB upsert. Emits BAGGAGE_STATUS_UPDATED to SSE stream.",
        "tags": ["Baggage Service"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["baggageId", "status"],
                "properties": {
                  "baggageId": { "type": "string", "example": "BAG-77102" },
                  "status": { "type": "string", "example": "LOADED_ON_AIRCRAFT" },
                  "location": { "type": "string", "example": "London (LHR)" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Status synchronized" },
          "400": { "description": "Invalid status value" }
        }
      }
    },
    "/api/checkin/process": {
      "post": {
        "summary": "Process passenger check-in and issue boarding pass",
        "description": "Orchestrates check-in: validates data, generates boarding pass with seat/gate, creates baggage manifest tokens. Emits PASSENGER_CHECKED_IN event.",
        "tags": ["Check-In Service"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["passengerName", "flightId"],
                "properties": {
                  "passengerName": { "type": "string", "example": "James Smith" },
                  "flightId": { "type": "string", "example": "AL-101" },
                  "baggageCount": { "type": "integer", "example": 2 },
                  "departureTime": { "type": "string", "format": "date-time" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Boarding pass issued", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BoardingPass" } } } },
          "400": { "description": "Missing required fields" }
        }
      }
    },
    "/api/checkin/status/{checkinId}": {
      "get": {
        "summary": "Get check-in record by ID",
        "tags": ["Check-In Service"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "checkinId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Check-in record" },
          "404": { "description": "Not found" }
        }
      }
    },
...
  }
}